move user fields logic into own contribution
