+++ arduino-ide-extension/src/browser/contributions/upload-sketch.ts
@@ -1,7 +1,7 @@
 import { inject, injectable } from '@theia/core/shared/inversify';
 import { Emitter } from '@theia/core/lib/common/event';
-import { BoardUserField, CoreService, Port } from '../../common/protocol';
-import { ArduinoMenus, PlaceholderMenuNode } from '../menu/arduino-menus';
+import { CoreService, Port } from '../../common/protocol';
+import { ArduinoMenus } from '../menu/arduino-menus';
 import { ArduinoToolbar } from '../toolbar/arduino-toolbar';
 import {
   Command,
@@ -11,93 +11,36 @@ import {
   TabBarToolbarRegistry,
   CoreServiceContribution,
 } from './contribution';
-import { UserFieldsDialog } from '../dialogs/user-fields/user-fields-dialog';
-import { deepClone, DisposableCollection, nls } from '@theia/core/lib/common';
+import { deepClone, nls } from '@theia/core/lib/common';
 import { CurrentSketch } from '../../common/protocol/sketches-service-client-impl';
 import type { VerifySketchParams } from './verify-sketch';
+import { UserFields } from './user-fields';
 
 @injectable()
 export class UploadSketch extends CoreServiceContribution {
-  @inject(MenuModelRegistry)
-  private readonly menuRegistry: MenuModelRegistry;
-
-  @inject(UserFieldsDialog)
-  private readonly userFieldsDialog: UserFieldsDialog;
-
-  private boardRequiresUserFields = false;
-  private userFieldsSet = false;
-  private readonly cachedUserFields: Map<string, BoardUserField[]> = new Map();
-  private readonly menuActionsDisposables = new DisposableCollection();
-
   private readonly onDidChangeEmitter = new Emitter<void>();
   private readonly onDidChange = this.onDidChangeEmitter.event;
   private uploadInProgress = false;
 
-  protected override init(): void {
-    super.init();
-    this.boardsServiceProvider.onBoardsConfigChanged(async () => {
-      const userFields =
-        await this.boardsServiceProvider.selectedBoardUserFields();
-      this.boardRequiresUserFields = userFields.length > 0;
-      this.registerMenus(this.menuRegistry);
-    });
-  }
-
-  private selectedFqbnAddress(): string {
-    const { boardsConfig } = this.boardsServiceProvider;
-    const fqbn = boardsConfig.selectedBoard?.fqbn;
-    if (!fqbn) {
-      return '';
-    }
-    const address =
-      boardsConfig.selectedBoard?.port?.address ||
-      boardsConfig.selectedPort?.address;
-    if (!address) {
-      return '';
-    }
-    return fqbn + '|' + address;
-  }
+  @inject(UserFields)
+  private readonly userFields: UserFields;
 
   override registerCommands(registry: CommandRegistry): void {
     registry.registerCommand(UploadSketch.Commands.UPLOAD_SKETCH, {
       execute: async () => {
-        const key = this.selectedFqbnAddress();
-        /*
-          If the board requires to be configured with user fields, we want
-          to show the user fields dialog, but if they weren't already
-          filled in or if they were filled in, but the previous upload failed.
-        */
-        if (
-          this.boardRequiresUserFields &&
-          key &&
-          (!this.cachedUserFields.has(key) || !this.userFieldsSet)
-        ) {
-          const userFieldsFilledIn = Boolean(
-            await this.showUserFieldsDialog(key)
-          );
-          if (!userFieldsFilledIn) {
-            return;
-          }
+        if (await this.userFields.checkUserFieldsDialog(false)) {
+          this.uploadSketch();
         }
-        this.uploadSketch();
       },
       isEnabled: () => !this.uploadInProgress,
     });
     registry.registerCommand(UploadSketch.Commands.UPLOAD_WITH_CONFIGURATION, {
       execute: async () => {
-        const key = this.selectedFqbnAddress();
-        if (!key) {
-          return;
+        if (await this.userFields.checkUserFieldsDialog(true)) {
+          this.uploadSketch();
         }
-        const userFieldsFilledIn = Boolean(
-          await this.showUserFieldsDialog(key)
-        );
-        if (!userFieldsFilledIn) {
-          return;
-        }
-        this.uploadSketch();
       },
-      isEnabled: () => !this.uploadInProgress && this.boardRequiresUserFields,
+      isEnabled: () => !this.uploadInProgress && this.userFields.isRequired(),
     });
     registry.registerCommand(
       UploadSketch.Commands.UPLOAD_SKETCH_USING_PROGRAMMER,
@@ -117,45 +60,20 @@ export class UploadSketch extends CoreServiceContribution {
   }
 
   override registerMenus(registry: MenuModelRegistry): void {
-    this.menuActionsDisposables.dispose();
-    this.menuActionsDisposables.push(
-      registry.registerMenuAction(ArduinoMenus.SKETCH__MAIN_GROUP, {
-        commandId: UploadSketch.Commands.UPLOAD_SKETCH.id,
-        label: nls.localize('arduino/sketch/upload', 'Upload'),
-        order: '1',
-      })
-    );
-    if (this.boardRequiresUserFields) {
-      this.menuActionsDisposables.push(
-        registry.registerMenuAction(ArduinoMenus.SKETCH__MAIN_GROUP, {
-          commandId: UploadSketch.Commands.UPLOAD_WITH_CONFIGURATION.id,
-          label: UploadSketch.Commands.UPLOAD_WITH_CONFIGURATION.label,
-          order: '2',
-        })
-      );
-    } else {
-      this.menuActionsDisposables.push(
-        registry.registerMenuNode(
-          ArduinoMenus.SKETCH__MAIN_GROUP,
-          new PlaceholderMenuNode(
-            ArduinoMenus.SKETCH__MAIN_GROUP,
-            // commandId: UploadSketch.Commands.UPLOAD_WITH_CONFIGURATION.id,
-            UploadSketch.Commands.UPLOAD_WITH_CONFIGURATION.label,
-            { order: '2' }
-          )
-        )
-      );
-    }
-    this.menuActionsDisposables.push(
-      registry.registerMenuAction(ArduinoMenus.SKETCH__MAIN_GROUP, {
-        commandId: UploadSketch.Commands.UPLOAD_SKETCH_USING_PROGRAMMER.id,
-        label: nls.localize(
-          'arduino/sketch/uploadUsingProgrammer',
-          'Upload Using Programmer'
-        ),
-        order: '3',
-      })
-    );
+    registry.registerMenuAction(ArduinoMenus.SKETCH__MAIN_GROUP, {
+      commandId: UploadSketch.Commands.UPLOAD_SKETCH.id,
+      label: nls.localize('arduino/sketch/upload', 'Upload'),
+      order: '1',
+    });
+
+    registry.registerMenuAction(ArduinoMenus.SKETCH__MAIN_GROUP, {
+      commandId: UploadSketch.Commands.UPLOAD_SKETCH_USING_PROGRAMMER.id,
+      label: nls.localize(
+        'arduino/sketch/uploadUsingProgrammer',
+        'Upload Using Programmer'
+      ),
+      order: '3',
+    });
   }
 
   override registerKeybindings(registry: KeybindingRegistry): void {
@@ -212,20 +130,8 @@ export class UploadSketch extends CoreServiceContribution {
         return;
       }
 
-      if (this.boardRequiresUserFields) {
-        // TODO: This does not belong here.
-        // IDE2 should not do any preliminary checks but let the CLI fail and then toast a user consumable error message.
-        if (uploadOptions.userFields.length === 0) {
-          this.messageService.error(
-            nls.localize(
-              'arduino/sketch/userFieldsNotFoundError',
-              "Can't find user fields for connected board"
-            )
-          );
-          this.userFieldsSet = false;
-          return;
-        }
-        this.userFieldsSet = true;
+      if (!this.userFields.checkUserFieldsForUpload()) {
+        return;
       }
 
       await this.doWithProgress({
@@ -240,13 +146,7 @@ export class UploadSketch extends CoreServiceContribution {
         { timeout: 3000 }
       );
     } catch (e) {
-      if (
-        this.boardRequiresUserFields &&
-        typeof e.message === 'string' &&
-        e.message.startsWith('Upload error:')
-      ) {
-        this.userFieldsSet = false;
-      }
+      this.userFields.notifyFailedWithError(e);
       this.handleError(e);
     } finally {
       this.uploadInProgress = false;
@@ -263,7 +163,7 @@ export class UploadSketch extends CoreServiceContribution {
     if (!CurrentSketch.isValid(sketch)) {
       return undefined;
     }
-    const userFields = this.userFields();
+    const userFields = this.userFields.getUserFields();
     const { boardsConfig } = this.boardsServiceProvider;
     const [fqbn, { selectedProgrammer: programmer }, verify, verbose] =
       await Promise.all([
@@ -306,10 +206,6 @@ export class UploadSketch extends CoreServiceContribution {
     return port;
   }
 
-  private userFields(): BoardUserField[] {
-    return this.cachedUserFields.get(this.selectedFqbnAddress()) ?? [];
-  }
-
   /**
    * Converts the `VENDOR:ARCHITECTURE:BOARD_ID[:MENU_ID=OPTION_ID[,MENU2_ID=OPTION_ID ...]]` FQBN to
    * `VENDOR:ARCHITECTURE:BOARD_ID` format.
@@ -321,23 +217,6 @@ export class UploadSketch extends CoreServiceContribution {
     }
     const [vendor, arch, id] = fqbn.split(':');
     return `${vendor}:${arch}:${id}`;
-  }
-
-  private async showUserFieldsDialog(
-    key: string
-  ): Promise<BoardUserField[] | undefined> {
-    const cached = this.cachedUserFields.get(key);
-    // Deep clone the array of board fields to avoid editing the cached ones
-    this.userFieldsDialog.value = (
-      cached ?? (await this.boardsServiceProvider.selectedBoardUserFields())
-    ).map((f) => ({ ...f }));
-    const result = await this.userFieldsDialog.open();
-    if (!result) {
-      return;
-    }
-    this.userFieldsSet = true;
-    this.cachedUserFields.set(key, result);
-    return result;
   }
 }
